fix cycleImport
